test(4259): extract post-loop guard coverage
Keep the save-store parent at its frozen test-residue ceiling by moving the post-loop identity mutation test into a focused child module without changing its assertions.
